refactor: move aggregation utilities to shared/ for cross-process reuse
