+++ main/collection-helpers.js
@@ -1,20 +1,10 @@
 /**
  * Generic collection utilities shared across helper modules.
+ *
+ * countBy is re-exported from shared/aggregation-utils.js
+ * for cross-process reuse (main + renderer).
  */
 
-/**
- * Counts occurrences of each key produced by keyFn.
- * @param {Array} items
- * @param {(item: unknown) => string} keyFn - Returns the key for each item
- * @returns {Record<string, number>} map of key -> count
- */
-function countBy(items, keyFn) {
-  const counts = {};
-  for (const item of items) {
-    const key = keyFn(item);
-    counts[key] = (counts[key] || 0) + 1;
-  }
-  return counts;
-}
+const { countBy } = require('../shared/aggregation-utils');
 
 module.exports = { countBy };
